pyproject.toml
[tool.poetry]
name = "yt-listener"
version = "0.1.0"
description = ""
authors = ["SirAbhi13 <abhinav.sny.2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.0"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
